% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_active_consults.R
\name{calc_active_consults}
\alias{calc_active_consults}
\title{Query um quadro de dados com consultas clínicas para pacientes activos em TARV}
\usage{
calc_active_consults(con, patient_uuids)
}
\arguments{
\item{con}{Ligação à base de dados MozART 2.0}

\item{patient_uuids}{Um vector contendo os patient_uuid unicos dos pacientes activos em TARV}
}
\value{
\code{calc_active_consults} devolve um quadro de dados com consultas clínicas para pacientes activos em TARV
}
\description{
Query um quadro de dados com consultas clínicas para pacientes activos em TARV
}
\examples{
\dontrun{

 df <- calc_active_consults(
          con,
          patient_uuids = tx_uuids)}
}
