% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_age_var.R
\name{calc_age_var}
\alias{calc_age_var}
\title{Calcular a idade do cliente em anos}
\usage{
calc_age_var(
  df,
  birth_date = "birthdate",
  ref_date = Sys.Date(),
  variable_name = "age"
)
}
\arguments{
\item{df}{Quadro de dados contendo a data de nasicmento do cliente}

\item{birth_date}{Coluna do quadro de dados utilizada para introduzir a data de nascimento do cliente (default = birthdate)}

\item{ref_date}{Data de referência para calcular a idade do cliente (default = Sys.Date())}

\item{variable_name}{Nome da variável que será criada no quadro de dados com a idade do cliente em anos (default = 'age')}
}
\value{
Quadro de dados com a idade do cliente em anos
}
\description{
Calcular a idade do cliente em anos
}
\examples{
\dontrun{

 df <- calc_age_var(df)}
}
