% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_arv_pickups.R
\name{calc_arv_pickups}
\alias{calc_arv_pickups}
\title{Query estatísticas anónimas para levantamentos TARV}
\usage{
calc_arv_pickups(
  con,
  enddate,
  arv_pickup_source = "FILA",
  n_months = 12,
  filter_by_location = FALSE,
  location_uuid = "4be5f1a9-832c-4717-be41-ef4b6311c0ef"
)
}
\arguments{
\item{con}{Ligação à base de dados MozART 2.0}

\item{enddate}{Data de fecho do período (introduzir como “AAAA-MM-DD”)}

\item{arv_pickup_source}{Fonte de levantamento de ARV. Opções são "FILA", "Reception", "Any". Default é "FILA"}

\item{n_months}{Número de meses para retroceder a partir da data de fecho do período (enddate). Default é 12 meses}

\item{filter_by_location}{Lógico. Se \code{TRUE}, o query corre sobre uma unidade sanitária específica.  Se \code{FALSE},  o query corre sobre todas as unidade sanitária contidas na base de dados}

\item{location_uuid}{location_uuid da unidade sanitária a filtrar quando filter_by_location é definido como \code{TRUE}}
}
\value{
Um quadro de dados contendo uma listagem individual dos levantamentos TARV
}
\description{
Query estatísticas anónimas para levantamentos TARV
}
\examples{
\dontrun{

 df <- calc_arv_pickups(
          con,
          enddate = '2024-09-20',
          n_months = 6,
          arv_pickup_source = "FILA",
          filter_by_location = TRUE,
          location_uuid = location_meripo)}
}
