% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_client_age.R
\name{calc_client_age}
\alias{calc_client_age}
\title{Calcular a idade do cliente em anos}
\usage{
calc_client_age(birth_date, ref_date = Sys.Date())
}
\arguments{
\item{birth_date}{Coluna do quadro de dados utilizada para introduzir a data de nascimento do cliente (default = birthdate)}

\item{ref_date}{Data de referência para calcular a idade do cliente (default = Sys.Date())}
}
\value{
Quadro de dados com a idade do cliente em anos
}
\description{
Calcular a idade do cliente em anos
}
\examples{
\dontrun{

 df <- calculate_age(df)}
}
