% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_active_consults.R, R/calc_consults.R
\name{calc_consults}
\alias{calc_consults}
\title{Query consultas clínicas realizadas}
\usage{
calc_consults(
  con,
  enddate = Sys.Date(),
  opendate = NULL,
  client_status = "All",
  defaulter_tolerance = 28,
  filter_by_location = FALSE,
  location_uuid = NULL
)

calc_consults(
  con,
  enddate = Sys.Date(),
  opendate = NULL,
  client_status = "All",
  defaulter_tolerance = 28,
  filter_by_location = FALSE,
  location_uuid = NULL
)
}
\arguments{
\item{con}{Ligação à base de dados MozART 2.0}

\item{enddate}{Data de fecho do período (introduzir como “AAAA-MM-DD”)}

\item{opendate}{Data de abertura do período (introduzir como “AAAA-MM-DD”)}

\item{client_status}{XXX}

\item{defaulter_tolerance}{Número de dias de tolerância antes dos faltosos serem considerados inactivos (abandonos)}

\item{filter_by_location}{Lógico. Se \code{TRUE}, o query corre sobre uma unidade sanitária específica.  Se \code{FALSE},  o query corre sobre todas as unidade sanitária contidas na base de dados}

\item{location_uuid}{location_uuid da unidade sanitária a filtrar quando filter_by_location é definido como \code{TRUE}}
}
\value{
\code{calc_consults} devolve um quadro de dados com consultas clínicas para pacientes activos em TARV

Um quadro de dados contendo uma listagem das consultas clínicas realizadas
}
\description{
Query consultas clínicas realizadas

Query consultas clínicas realizadas
}
\examples{
\dontrun{

 df <- calc_consults(
          con,
          patient_uuids = tx_uuids)}
\dontrun{

 df <- calc_consults(
          con,
          patient_uuids = tx_uuids)}
}
