% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_tx_new.R
\name{calc_tx_new}
\alias{calc_tx_new}
\title{Calculate line-level listing for patients newly initiated on ART}
\usage{
calc_tx_new(
  con,
  opendate,
  enddate,
  filter_by_location = FALSE,
  location_uuid = "4be5f1a9-832c-4717-be41-ef4b6311c0ef"
)
}
\arguments{
\item{con}{MozART 2.0 database connection}

\item{opendate}{Open date for the reporting period for which new ART initiations are calculated}

\item{enddate}{Close date for the reporting period for which new ART initiations are calculated}

\item{filter_by_location}{Logical indicating whether to filter query to a specific location (health facility) or not}

\item{location_uuid}{Location_uuid of the health facility to filter by when filter_by_location is TRUE}
}
\value{
A data frame containing line-level listing of patients newly initiated on ART
}
\description{
Calculate line-level listing for patients newly initiated on ART
}
\examples{
\dontrun{

 df <- calc_tx_new()}
}
