% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_age.R
\name{calculate_age}
\alias{calculate_age}
\title{Calculate client age in years}
\usage{
calculate_age(birth_date, ref_date = Sys.Date())
}
\arguments{
\item{birth_date}{Coluna do quadro de dados utilizada para introduzir a data de nascimento do cliente (default = birthdate)}

\item{ref_date}{Data de referência para calcular a idade do cliente (default = Sys.Date())}
}
\value{
Quadro de dados com a idade do cliente em anos
}
\description{
Calculate client age in years
}
\examples{
\dontrun{

 df <- calculate_age(df)}
}
