% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_location_lookup}
\alias{data_location_lookup}
\title{Tabela de referência: location_uuid}
\format{
Quadro de dados com variáveis:
\describe{
\item{location_uuid}{código único de identificação do local}
\item{data_location_lookup}{datim_id}
\item{data_location_lookup}{sisma_id}
\item{data_location_lookup}{province}
\item{data_location_lookup}{district}
\item{data_location_lookup}{datim_name}
\item{data_location_lookup}{mozart_name}
}
}
\usage{
data(data_location_lookup)
}
\description{
Source: \href{https://github.com/usaid-mozambique/sonata/blob/main/R/data.R}{\code{R/data.R}}.
}
\keyword{datasets}
