% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_observation_lookup}
\alias{data_observation_lookup}
\title{Referência: data_observation_lookup}
\format{
Quadro de dados com variáveis:
\describe{
\item{concept_id}{Identificador do conceito}
\item{concept_name}{Nome do conceito}
}
}
\usage{
data(data_observation_lookup)
}
\description{
Source: \href{https://github.com/usaid-mozambique/sonata/blob/main/R/data.R}{\code{R/data.R}}.
}
\keyword{datasets}
