% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duck_utilities.R
\name{duck_connection}
\alias{duck_connection}
\title{Initiate duckdb connection}
\usage{
duck_connection(dbfile = ":memory:", ...)
}
\arguments{
\item{dbfile}{database file}

\item{...}{test fix}
}
\description{
Initiate duckdb connection
}
