% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_account}
\alias{get_account}
\title{Get account details}
\usage{
get_account(name)
}
\arguments{
\item{name}{Service name of the account}
}
\value{
key / value pair as list containing details of the account (invisible)
}
\description{
Get account details
}
\note{
Inspired by \code{grabr::lazy_secrets()}
}
\examples{
\dontrun{
get_account(name = 's3')}

}
\seealso{
Other authentication: 
\code{\link{set_account}()}
}
\concept{authentication}
