% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{mysql_connection}
\alias{mysql_connection}
\title{Ligar ao MozART 2.0}
\usage{
mysql_connection(
  db_name,
  db_user,
  db_pass,
  db_host = "localhost",
  db_port = 3306
)
}
\arguments{
\item{db_name}{Nome da base de dados}

\item{db_user}{Utilizador da base de dados}

\item{db_pass}{Senha da base de dados}

\item{db_host}{Host da base de dados}

\item{db_port}{Porta da base de dados}
}
\value{
Uma conexão à base de dados
}
\description{
Ligar ao MozART 2.0
}
\examples{
\dontrun{

 df <- mysql_connection()}
}
