% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode.R
\name{process_recoding}
\alias{process_recoding}
\title{Recodificar múltiplas variáveis definidas pelo utilizador}
\usage{
process_recoding(
  df,
  columns_to_recode,
  options = list(df_disp_mode = data_type_id_lookup, df_regimen = data_type_id_lookup,
    df_form = data_type_id_lookup, df_location = data_location_lookup, age_column =
    "age", sex_column = "gender")
)
}
\arguments{
\item{df}{Quadro de dados contendo a variável a recodificar}

\item{columns_to_recode}{Vector de strings contendo os nomes das colunas a recodificar}

\item{options}{Lista de parâmetros opcionais para personalizar a recodificação}
}
\value{
\code{process_recoding} devolve um quadro de dados com colunas recodificadas
}
\description{
Recodificar múltiplas variáveis definidas pelo utilizador
}
