% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_age_out.R
\name{project_age_out}
\alias{project_age_out}
\title{Detectar clients com transição etária}
\usage{
project_age_out(df, ref_date)
}
\arguments{
\item{df}{Quadro de dados contendo a data de nasicmento do cliente}

\item{ref_date}{Data de referência para calcular a idade projectada do cliente}
}
\value{
Quadro de dados com variáveis idade, idade projectada, faixa etária, faixa etária projectada, e colunas que indicam casos de transição etária
}
\description{
Detectar clients com transição etária
}
\examples{
\dontrun{

 df <- project_age_out(df)}
}
