% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_active_uuid.R
\name{pull_active_uuid}
\alias{pull_active_uuid}
\title{Query uuids dos activos no TARV}
\usage{
pull_active_uuid(
  con,
  enddate,
  defaulter_tolerance = 28,
  filter_by_location = TRUE,
  location_uuid = "e5f01eee-2392-49b4-a5bf-5cf593fc8f21"
)
}
\arguments{
\item{con}{Ligação à base de dados MozART 2.0}

\item{enddate}{Data de fecho do período (introduzir como “AAAA-MM-DD”)}

\item{defaulter_tolerance}{Número de dias de tolerância antes dos faltosos serem considerados inactivos (abandonos)}

\item{filter_by_location}{Lógico. Se \code{TRUE}, o query corre sobre uma unidade sanitária específica.  Se \code{FALSE},  o query corre sobre todas as unidade sanitária contidas na base de dados}

\item{location_uuid}{location_uuid da unidade sanitária a filtrar quando filter_by_location é definido como \code{TRUE}}
}
\value{
\code{pull_active_uuid} devolve um vector contendo os patient_uuid unicos dos pacientes activos em TARV
}
\description{
\code{pull_active_uuid()} devolve um vector dos uuid's dos clientes activos em TARV a partir de uma data definida pelo utilizador
}
\examples{
\dontrun{

#' Caso de uso simples executado sobre uma US, tolerância de 28 dias
 df <- pull_active_uuid(
          con,
          enddate = '2024-09-20',
          defaulter_tolerance = 28)}
}
