% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_client_uuid.R
\name{pull_client_uuid}
\alias{pull_client_uuid}
\title{Query uuids para activos ou novos inícios ao TARV}
\usage{
pull_client_uuid(
  con,
  opendate = NULL,
  enddate = Sys.Date(),
  client_status = "Active",
  defaulter_tolerance = 28,
  filter_by_location = TRUE,
  location_uuid = "e5f01eee-2392-49b4-a5bf-5cf593fc8f21"
)
}
\arguments{
\item{con}{Ligação à base de dados MozART 2.0}

\item{opendate}{Data de abertura do período (introduzir como “AAAA-MM-DD”)}

\item{enddate}{Data de fecho do período (introduzir como “AAAA-MM-DD”)}

\item{client_status}{Character. Defines which UUIDs to pull. Options: \code{"Active"} (default)
to retrieve active clients, \code{"New"} to retrieve new clients.}

\item{defaulter_tolerance}{Numeric. Number of days used to define a defaulter
when \code{client_status = "Active"}. Defaults to 28 days.}

\item{filter_by_location}{Se \code{TRUE}, o query corre sobre uma unidade sanitária específica.  Se \code{FALSE}, o query corre sobre todas as unidade sanitária contidas no MozART 2.0}

\item{location_uuid}{location_uuid da unidade sanitária a filtrar quando filter_by_location é definido como \code{TRUE}}
}
\value{
Um vector de UUIDs de clientes que correspondem aos critérios especificados.
}
\description{
Esta função funciona como um invólucro para devolver UUIDs de clientes com base no seu
estado (“Active” ou “New”). Ela chama \code{pull_active_uuid()} para clientes ativos ou
\code{pull_new_uuid()} para novos clientes, dependendo do argumento \code{client_status}.
}
\examples{
\dontrun{
# Caso de uso simples executado sobre todas US
active_uuids <- pull_client_uuid(con,
                                 enddate = "2024-09-20",
                                 filter_by_location = FALSE)

# Devolver novos clientes a partir de uma data de abertura especificada
new_uuids <- pull_client_uuid(con,
                              opendate = "2024-01-01",
                              enddate = "2024-09-20",
                              client_status = "New")

# Devolver clientes activos para uma US específica com uma tolerância de 59 dias
active_uuids <- pull_client_uuid(con,
                                 enddate = "2024-09-20",
                                 defaulter_tolerance = 59,
                                 location_uuid = "custom-location-uuid")}
}
