% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_new_uuid.R
\name{pull_new_uuid}
\alias{pull_new_uuid}
\title{Query uuids dos novos inícios ao TARV}
\usage{
pull_new_uuid(
  con,
  opendate,
  enddate,
  filter_by_location = TRUE,
  location_uuid = "e5f01eee-2392-49b4-a5bf-5cf593fc8f21"
)
}
\arguments{
\item{con}{Ligação à base de dados MozART 2.0}

\item{opendate}{Data de abertura do período (introduzir como “AAAA-MM-DD”)}

\item{enddate}{Data de fecho do período (introduzir como “AAAA-MM-DD”)}

\item{filter_by_location}{Se \code{TRUE}, o query corre sobre uma unidade sanitária específica.  Se \code{FALSE}, o query corre sobre todas as unidade sanitária contidas no MozART 2.0}

\item{location_uuid}{location_uuid da unidade sanitária a filtrar quando filter_by_location é definido como \code{TRUE}}
}
\value{
\code{pull_new_uuid} devolve um vector contendo os patient_uuid unicos dos pacientes activos em TARV
}
\description{
\code{pull_new_uuid()} devolve um vector dos uuid's dos clientes activos em TARV a partir de uma data definida pelo utilizador
}
\examples{
\dontrun{

# Caso de uso simples executado sobre todas US
uuids <- pull_new_uuid(con, opendate = "2024-06-21", enddate = "2024-09-20", filter_by_location = FALSE)

# # Caso de uso simples executado sobre uma US
uuids <- pull_new_uuid(con, opendate = "2024-06-21", enddate = "2024-09-20",
                       filter_by_location = TRUE, location_uuid = 'e3eb1a1b-be07-4af2-9360-5d7046910576')
}
}
