% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode.R
\name{recode_age}
\alias{recode_age}
\title{Recodificar idade}
\usage{
recode_age(df, age_column = "age")
}
\arguments{
\item{df}{Quadro de dados contendo a variável a recodificar}

\item{age_column}{Variável no quadro de dados introduzido contendo a idade do cliente}
}
\value{
\code{recode_age} devolve um quadro de dados com a coluna idade recodificada
}
\description{
Recodificar idade
}
\examples{
\dontrun{

 df <- recode_age(
          df = df,
          age_column = "age"))}
}
