% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode.R
\name{recode_disp_mode}
\alias{recode_disp_mode}
\title{Recodificar modo de dispensa TARV}
\usage{
recode_disp_mode(df, df_disp_mode = data_type_id_lookup, keep_id = FALSE)
}
\arguments{
\item{df}{Quadro de dados contendo a variável a recodificar}

\item{df_disp_mode}{Objecto de tabela de pesquisa usado para recodificação}

\item{keep_id}{Manter a coluna regimen_id após a recodificação (Logical T/F)}
}
\value{
\code{recode_disp_mode} devolve um quadro de dados com a coluna regimen recodificada
}
\description{
Recodificar modo de dispensa TARV
}
\examples{
\dontrun{

 df <- recode_disp_mode(
          df = df,
          df_disp_mode = data_type_id_lookup,
          keep_id = FALSE)}
}
