% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode.R
\name{recode_form}
\alias{recode_form}
\title{Recodificar formulario primario}
\usage{
recode_form(df, df_form = data_type_id_lookup, keep_id = FALSE)
}
\arguments{
\item{df}{Quadro de dados contendo a variável a recodificar}

\item{df_form}{Objecto de tabela de pesquisa usado para recodificação}

\item{keep_id}{Manter a coluna regimen_id após a recodificação (Logical T/F)}
}
\value{
\code{recode_form} devolve um quadro de dados com a coluna regimen recodificada
}
\description{
Recodificar formulario primario
}
\examples{
\dontrun{

 df <- recode_form(
          df = df,
          df_form = data_type_id_lookup,
          keep_id = FALSE))}
}
