% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode.R
\name{recode_location}
\alias{recode_location}
\title{Recodificar location_uuid's}
\usage{
recode_location(df, df_location = data_location_lookup)
}
\arguments{
\item{df}{Quadro de dados contendo a variável a recodificar}

\item{df_location}{Objecto de tabela de pesquisa usado para recodificação}
}
\value{
\code{recode_location} devolve um quadro de dados com a coluna location_uuid recodificada
}
\description{
Recodificar location_uuid's
}
\examples{
\dontrun{

 df <- recode_location(
          df = df,
          df_location = data_location_lookup)}
}
