% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode.R
\name{recode_regimen}
\alias{recode_regimen}
\title{Recodificar regimen TARV}
\usage{
recode_regimen(df, df_regimen = data_type_id_lookup, keep_id = FALSE)
}
\arguments{
\item{df}{Quadro de dados contendo a variável a recodificar}

\item{df_regimen}{Objecto de tabela de pesquisa usado para recodificação}

\item{keep_id}{Manter a coluna regimen_id após a recodificação (Logical T/F)}
}
\value{
\code{recode_regimen} devolve um quadro de dados com a coluna regimen recodificada
}
\description{
Recodificar regimen TARV
}
\examples{
\dontrun{

 df <- recode_regimen(
          df = df,
          df_regimen = data_type_id_lookup,
          keep_id = FALSE)}
}
