% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode.R
\name{recode_sex}
\alias{recode_sex}
\title{Recodificar sexo}
\usage{
recode_sex(df, sex_column = "gender")
}
\arguments{
\item{df}{Quadro de dados contendo a variável a recodificar}

\item{sex_column}{Variável no quadro de dados introduzido contendo o sexo do cliente}
}
\value{
\code{recode_sex} devolve um quadro de dados com a coluna sexo recodificada
}
\description{
Recodificar sexo
}
\examples{
\dontrun{

 df <- recode_sex(
          df = df,
          sex_column = "sex"))}
}
