% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode.R
\name{recode_state}
\alias{recode_state}
\title{Recodificar estado do cliente}
\usage{
recode_state(df, df_state = data_type_id_lookup, keep_id = FALSE)
}
\arguments{
\item{df}{Quadro de dados contendo a variável a recodificar}

\item{df_state}{Objecto de tabela de pesquisa usado para recodificação}

\item{keep_id}{Manter a coluna state_id após a recodificação (Logical T/F)}
}
\value{
\code{recode_state} devolve um quadro de dados com a coluna state recodificada
}
\description{
\code{recode_state()} devolve uma apresentação mais compreensível do estado do cliente
}
\examples{
\dontrun{

 df <- recode_state(
          df = df,
          df_state = data_type_id_lookup,
          keep_id = FALSE)}
}
