% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{set_account}
\alias{set_account}
\title{Create / Update account}
\usage{
set_account(name, keys = c("username", "password"), update = FALSE)
}
\arguments{
\item{name}{Service name of the account}

\item{keys}{List of account key names}

\item{update}{Should an existing account be overwriten}
}
\description{
Create / Update account
}
\examples{
\dontrun{
set_account(name = 's3', keys = c("access", "secret"))}

}
\seealso{
Other authentication: 
\code{\link{get_account}()}
}
\concept{authentication}
