% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_arv_pickups.R
\name{calc_arv_pickups}
\alias{calc_arv_pickups}
\title{Query levantamentos TARV}
\usage{
calc_arv_pickups(
  con,
  enddate,
  opendate,
  arv_pickup_source = "FILA",
  filter_by_location = TRUE,
  location_uuid = "e5f01eee-2392-49b4-a5bf-5cf593fc8f21"
)
}
\arguments{
\item{con}{Ligação à base de dados MozART 2.0}

\item{enddate}{Data do fecho do período (introduzir como “AAAA-MM-DD”)}

\item{opendate}{Data da abertura do período (introduzir como “AAAA-MM-DD”)}

\item{arv_pickup_source}{Fonte de levantamento de ARV. Opções são "FILA", "Reception", "Any". Default é "FILA
\itemize{
\item \code{"FILA"} por defeito, devolve levantamentos documentados na FILA ou fonte alternativa da farmácia
\item \code{"Reception"} devolve levantamentos documentados na Ficha de Seguimento, Mapa dos Levantamentos
\item \code{"Any"} devolve todos levantamentos
}}

\item{filter_by_location}{\code{FALSE} por defeito
\itemize{
\item \code{TRUE} executa sobre uma US específica
\item \code{FALSE} executa sobre todas as US contidas no MozART 2.0
}}

\item{location_uuid}{location_uuid da US a filtrar quando filter_by_location é definido como \code{TRUE}}
}
\value{
Um quadro de dados contendo uma listagem individual dos levantamentos TARV
}
\description{
\code{calc_arv_pickups()} devolve uma listagem dos levantamentos ARV feito during um período especificado pelo utilizador
}
\examples{
\dontrun{

 df <- calc_arv_pickups(
          con
          opendate = '2024-06-20',
          enddate = '2024-09-20',
          arv_pickup_source = "FILA",
          filter_by_location = TRUE,
          location_uuid = location_meripo)}
}
