% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_consults.R
\name{calc_consults}
\alias{calc_consults}
\title{Query consultas clínicas HIV}
\usage{
calc_consults(
  con,
  client_status = "All",
  defaulter_tolerance = 28,
  opendate = NULL,
  enddate = Sys.Date(),
  filter_by_location = FALSE,
  location_uuid = NULL
)
}
\arguments{
\item{con}{Ligação à base de dados MozART 2.0}

\item{client_status}{Estado do cliente em TARV
\itemize{
\item \code{"All"}, por defeito, devolve todos clientes tendo realizado consultas clínicas
\item \code{"Active"}, devolve apenas os clientes activos em TARV
}}

\item{defaulter_tolerance}{Número de dias de tolerância antes dos faltosos serem considerados inactivos (abandonos)
\itemize{
\item \code{"28"} por defeito
}}

\item{opendate}{Data da abertura do período (introduzir como “AAAA-MM-DD”)}

\item{enddate}{Data do fecho do período usado para avialar o estado activo do client
\itemize{
\item \code{"Sys.Date"} Por defeito
\item Data customizada em formato (introduzir como “AAAA-MM-DD”)
\item Número customizado (por exemplo \code{"59"} na definição de MISAU)
}}

\item{filter_by_location}{Se \code{TRUE}, o query corre sobre uma unidade sanitária específica.  Se \code{FALSE}, o query corre sobre todas as unidade sanitária contidas no MozART 2.0}

\item{location_uuid}{location_uuid da unidade sanitária a filtrar quando filter_by_location é definido como \code{TRUE}}
}
\value{
Um quadro de dados contendo uma listagem das consultas clínicas HIV realizadas
}
\description{
\code{calc_consults()} executa um query que devolve uma listagem das consultas clínicas do HIV realizadas por um grupo de clientes definido pelo utilizador (activos em TARV ou todos clientes)
}
\examples{
\dontrun{

 # Caso de uso simples executado sobre todas US
 df <- calc_consults(con)

 # Caso de uso de activos, tolerância 59 dias, executado sobre uma US
 df <- calc_consults(
          con,
          enddate = '2024-09-20',
          client_status = "Active",
          defaulter_tolerance = 59,
          filter_by_location = TRUE,
          location_uuid = 'e3eb1a1b-be07-4af2-9360-5d7046910576')

 # Caso de uso de activos, limitando as consultas a 2022 para frente
 df <- calc_consults(
          con,
          client_status = "Active",
          enddate = '2024-09-20',
          opendate = '2022-01-10')}
}
