% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_tx_active.R
\name{calc_tx_active}
\alias{calc_tx_active}
\title{Query clientes activos em TARV}
\usage{
calc_tx_active(
  con,
  enddate,
  defaulter_tolerance = 28,
  filter_by_location = FALSE,
  location_uuid = "4be5f1a9-832c-4717-be41-ef4b6311c0ef"
)
}
\arguments{
\item{con}{Ligação à base de dados MozART 2.0}

\item{enddate}{Data de fecho do período (introduzir como “AAAA-MM-DD”)}

\item{defaulter_tolerance}{Número de dias de tolerância antes dos faltosos serem considerados inactivos (abandonos)}

\item{filter_by_location}{Lógico. Se \code{TRUE}, o query corre sobre uma unidade sanitária específica.  Se \code{FALSE},  o query corre sobre todas as unidade sanitária contidas na base de dados}

\item{location_uuid}{location_uuid da unidade sanitária a filtrar quando filter_by_location é definido como \code{TRUE}}
}
\value{
Um quadro de dados contendo uma listagem individual dos pacientes activos em TARV
}
\description{
\code{calc_tx_active()} devolve uma listagem dos clientes activos em TARV a partir de uma data definida pelo utilizador
}
\examples{
\dontrun{

#' Caso de uso simples executado sobre uma US, tolerância de 28 dias
df <- calc_tx_active(con,
                     enddate = '2024-09-20',
                     defaulter_tolerance = 28,
                     filter_by_location = TRUE,
                     location_uuid = 'e3eb1a1b-be07-4af2-9360-5d7046910576')

#' Caso de uso simples executado paras todas US, tolerância de 59 dias
df <- calc_tx_active(con,
                     enddate = '2024-09-20',
                     defaulter_tolerance = 59,
                     filter_by_location = FALSE)}
}
