% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_tx_active.R
\name{calc_tx_active}
\alias{calc_tx_active}
\title{Query anonymous statistics for Active on ART}
\usage{
calc_tx_active(
  con,
  enddate,
  defaulter_tolerance = 28,
  filter_by_location = FALSE,
  location_uuid = "4be5f1a9-832c-4717-be41-ef4b6311c0ef"
)
}
\arguments{
\item{con}{MozART 2.0 database connection}

\item{enddate}{Period close date (input as 'YYYY-MM-DD')}

\item{defaulter_tolerance}{Number of tolerance days before defaulters are considered inactive}

\item{filter_by_location}{Logical. If \code{TRUE}, filters query to a specific location.  If \code{FALSE} queries for all locations.}

\item{location_uuid}{Health facility location_uuid to filter by when filter_by_location is \code{TRUE}}
}
\value{
A data frame containing line-level listing of patients active on ART
}
\description{
Query anonymous statistics for Active on ART
}
\examples{
\dontrun{

 df <- calc_tx_active(
          con,
          enddate = '2024-09-20',
          defaulter_tolerance = 28,
          filter_by_location = TRUE,
          location_uuid = location_meripo)}
}
