% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_tx_new.R
\name{calc_tx_new}
\alias{calc_tx_new}
\title{Query anonymous statistics for Newly initiated on ART}
\usage{
calc_tx_new(
  con,
  opendate,
  enddate,
  filter_by_location = FALSE,
  location_uuid = "4be5f1a9-832c-4717-be41-ef4b6311c0ef"
)
}
\arguments{
\item{con}{MozART 2.0 database connection}

\item{opendate}{Period open date (input as 'YYYY-MM-DD')}

\item{enddate}{Period close date (input as 'YYYY-MM-DD')}

\item{filter_by_location}{Logical. If \code{TRUE}, filters query to a specific location.  If \code{FALSE} queries for all locations.}

\item{location_uuid}{Health facility location_uuid to filter by when filter_by_location is \code{TRUE}}
}
\value{
A data frame containing line-level listing of patients newly initiated on ART
}
\description{
Query anonymous statistics for Newly initiated on ART
}
\examples{
\dontrun{

 df <- calc_tx_new(
          con,
          opendate = "2024-06-21",
          enddate = "2024-09-20",
          filter_by_location = TRUE,
          location_uuid = location_meripo)}
}
