% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_location_lookup}
\alias{data_location_lookup}
\title{Referência: location_uuid}
\format{
Quadro de dados com variáveis:
\describe{
\item{location_uuid}{código único de identificação do local}
\item{datim_id}{Código único de identificação do Datim do local}
\item{sisma_id}{Código único de identificação do SISMA do local}
\item{province}{Província onde se situa o local}
\item{district}{Distrito em que o local está situado}
\item{datim_name}{Nome do local no Datim}
\item{mozart_name}{Nome do local no MozART}
}
}
\usage{
data(data_location_lookup)
}
\description{
Source: \href{https://github.com/usaid-mozambique/sonata/blob/main/R/data.R}{\code{R/data.R}}.
}
\keyword{datasets}
