% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_location_lookup}
\alias{data_location_lookup}
\title{Location Lookup Table}
\format{
A list of the variables included in the site list
\describe{
\item{data_location_lookup}{location_uuid}
\item{data_location_lookup}{datim_id}
\item{data_location_lookup}{sisma_id}
\item{data_location_lookup}{province}
\item{data_location_lookup}{district}
\item{data_location_lookup}{datim_name}
\item{data_location_lookup}{mozart_name}
}
}
\usage{
data(data_location_lookup)
}
\description{
List of indicators
}
\keyword{datasets}
