% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_observation_lookup}
\alias{data_observation_lookup}
\title{Observation Lookup Table}
\format{
A list of the variables included in the site list
\describe{
\item{data_observation_lookup}{concept_id}
\item{data_observation_lookup}{concept_name}
}
}
\usage{
data(data_observation_lookup)
}
\description{
List of indicators
}
\keyword{datasets}
