% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_type_id_lookup}
\alias{data_type_id_lookup}
\title{data_type_id_lookup}
\format{
Quadro de dados com variáveis:
\describe{
\item{data_type_id_lookup}{id}
\item{data_type_id_lookup}{table_name}
\item{data_type_id_lookup}{column_name}
\item{data_type_id_lookup}{id_type_lookup}
\item{data_type_id_lookup}{id_type_desc}
\item{data_type_id_lookup}{id_type_lookup_pt}
\item{data_type_id_lookup}{notes}
}
}
\usage{
data(data_type_id_lookup)
}
\description{
List of indicators
}
\keyword{datasets}
