% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{mysql_connection}
\alias{mysql_connection}
\title{Ligar ao MozART 2.0}
\usage{
mysql_connection(
  db_name,
  db_user,
  db_pass,
  db_host = "localhost",
  db_port = 3306
)
}
\arguments{
\item{db_name}{Nome da base de dados}

\item{db_user}{Utilizador da base de dados}

\item{db_pass}{Senha da base de dados}

\item{db_host}{Host da base de dados}

\item{db_port}{Porta da base de dados}
}
\value{
Uma conexão ao MozART 2.0
}
\description{
\code{mysql_connection()} cria uma ligação com o MozART 2.0 com base num conjunto de credenciais armazenadas de forma segura e definidas pelo utilizador
}
\examples{
\dontrun{

con <- mysql_connection(db_name = acct_db$dbname,
                        db_user = acct_db$username,
                        db_pass = acct_db$password,
                        db_host = acct_db$host,
                        db_port = acct_db$port)}
}
