% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_age_out.R
\name{project_age_out}
\alias{project_age_out}
\title{Detectar transição etária}
\usage{
project_age_out(df, ref_date)
}
\arguments{
\item{df}{Quadro de dados contendo a data de nasicmento do cliente}

\item{ref_date}{Data de referência para calcular a idade projectada do cliente}
}
\value{
Quadro de dados com variáveis idade, idade projectada, faixa etária, faixa etária projectada, e colunas que indicam casos de transição etária
}
\description{
\code{project_age_out()} devolve uma listagem dos clientes clientes com idades projectadas para uma data definida pelo utilizador, bem como colunas para indicar se fizeram a transição da idade pediátrica para a idade adulta e se a sua faixa etária de 5 anos mudou
}
\examples{
\dontrun{

#' Caso de uso simples projectando idades ate o início do ano 2026
 df <- project_age_out(df
                       ref_date = "2026-01-01")}
}
