% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_active_uuid.R
\name{pull_active_uuid}
\alias{pull_active_uuid}
\title{Query vector de patient_uuids de activos no TARV}
\usage{
pull_active_uuid(
  con,
  enddate,
  defaulter_tolerance = 28,
  filter_by_location = FALSE,
  location_uuid = "4be5f1a9-832c-4717-be41-ef4b6311c0ef"
)
}
\arguments{
\item{con}{Ligação à base de dados MozART 2.0}

\item{enddate}{Data de fecho do período (introduzir como “AAAA-MM-DD”)}

\item{defaulter_tolerance}{Número de dias de tolerância antes dos faltosos serem considerados inactivos (abandonos)}

\item{filter_by_location}{Lógico. Se \code{TRUE}, o query corre sobre uma unidade sanitária específica.  Se \code{FALSE},  o query corre sobre todas as unidade sanitária contidas na base de dados}

\item{location_uuid}{location_uuid da unidade sanitária a filtrar quando filter_by_location é definido como \code{TRUE}}
}
\value{
\code{pull_active_uuid} devolve um vector contendo os patient_uuid unicos dos pacientes activos em TARV
}
\description{
Query vector de patient_uuids de activos no TARV
}
\examples{
\dontrun{

 df <- pull_active_uuid(
          con,
          enddate = '2024-09-20',
          defaulter_tolerance = 28,
          filter_by_location = TRUE,
          location_uuid = location_meripo)}
}
