% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode.R
\name{recode_age}
\alias{recode_age}
\title{Recodificar idade}
\usage{
recode_age(df, age_column = "age", variable_name = "age_band")
}
\arguments{
\item{df}{Quadro de dados contendo a variável a recodificar}

\item{age_column}{Variável no quadro de dados introduzido contendo a idade do cliente}

\item{variable_name}{Nome da variável que será criada no quadro de dados com a faixa etária do cliente (default = 'age_band')}
}
\value{
\code{recode_age} devolve um quadro de dados com a coluna idade recodificada
}
\description{
\code{recode_age()} devolve a idade e a faixa etária do cliente
}
\examples{
\dontrun{

 df <- recode_age(
          df = df,
          age_column = "age")}
}
