% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode.R
\name{recode_cols}
\alias{recode_cols}
\title{Recodificar colunas}
\usage{
recode_cols(
  df,
  cols = c("mode_dispensation_id", "regimen_id", "location_uuid", "age", "form_id",
    "gender"),
  options = list(df_disp_mode = data_type_id_lookup, df_regimen = data_type_id_lookup,
    df_form = data_type_id_lookup, df_location = data_location_lookup, age_column =
    "age", sex_column = "gender")
)
}
\arguments{
\item{df}{Quadro de dados contendo colunas a serem recodificadas}

\item{cols}{Vector de strings contendo os nomes das colunas a recodificar}

\item{options}{Lista de parâmetros opcionais para personalizar a recodificação}
}
\value{
\code{recode_cols} devolve um quadro de dados com colunas recodificadas
}
\description{
\code{recode_cols} reatribui valores para colunas codificadas em quadros de dados MozART 2.0 comuns
}
\examples{
\dontrun{
 # Caso de uso para recodificar dispensation_id, regimen_id e location_uuid
 df <- recode_cols(
          df = df,
          cols = c("mode_dispensation_id",
                   "regimen_id",
                   "location_uuid"))}
}
